refactor(services): PR-3.7 queue_io + task_snapshot 顶层搬入 services/
  studio/queue_io.py      → studio/services/queue_io.py
  studio/task_snapshot.py → studio/services/task_snapshot.py

修复发现：
- queue_io 老 from . import db, presets_io（顶层）需改 from .. import db
- queue_io 用 presets_io.PresetError 需绕 presets 包入口直接 from .presets import io as presets_io

回归：1637 passed
